Fix problems found building on windows for python 3.6 Need to include <algorithm> for std::min and std::max. Need to define ssize_t that is Posix only. Only defined emacs_user and emacs_library once Cannot make Py::Long from EmacsChatr_t on windows - use a heavy handed cast.

+++ Editor/Source/pybemacs/pybemacs.cpp
@@ -639,7 +639,8 @@ public:
         static std::string fn_name( "termDisplayActivity" );
         try
         {
-            callOnSelf( fn_name, Py::Long( ch ) );
+            // QQQ: Windows: Why is the long cast needed?
+            callOnSelf( fn_name, Py::Long( long( ch ) ) );
         }
         catch( Py::Exception &e )
         {
